pyproject.toml
[project]
name = "adk-docker-uv"
version = "0.1.0"
description = "ADK on Docker, optimized with uv"
readme = "README.md"
authors = [
    { name = "doug", email = "110487462+doughayden@users.noreply.github.com" }
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "google-adk==1.18.0",
    "pydantic>=2.11.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
server = "adk_docker_uv.server:main"

[dependency-groups]
dev = [
    "ruff>=0.12.7,<1.0.0",
    "mypy>=1.17.1,<2.0.0",
    "pytest>=8.3.4,<9.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-asyncio>=1.2.0,<2.0.0",
    "ipykernel>=7.0.0,<8.0.0",
]

[build-system]
requires = ["uv_build>=0.8.21,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
src = ["src"]
target-version = "py313"
line-length = 88
fix = true
extend-exclude = ["notebooks"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # Allow assert statements in tests

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
packages = ["adk_docker_uv"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/adk_docker_uv/server.py",
    "src/adk_docker_uv/agent/agent.py",
    "src/adk_docker_uv/agent/prompt.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2
fail_under = 100

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince212:google.genai.types",
]
